+++ ios/Flutter/flutter_export_environment.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
 # This is a generated file; do not edit or check into version control.
 export "FLUTTER_ROOT=C:\Users\CESARE\flutter"
-export "FLUTTER_APPLICATION_PATH=C:\Users\CESARE\source\repository\package\auto_form"
+export "FLUTTER_APPLICATION_PATH=C:\Users\CESARE\source\repository\package\easy_auto_form"
 export "COCOAPODS_PARALLEL_CODE_SIGN=true"
 export "FLUTTER_TARGET=lib\main.dart"
 export "FLUTTER_BUILD_DIR=build"